Move SparklerConfig to config paackage
Now we have a bunch of classes related to config and
we created a package to group all these classes,
so SparklerConfig is better suited there
