+++ sparkler-plugins/urlfilter-regex/src/main/java/edu/usc/irds/sparkler/config/UrlFilterProps.java
@@ -2,7 +2,6 @@ package edu.usc.irds.sparkler.config;
 
 
 import edu.usc.irds.sparkler.BaseConfig;
-import edu.usc.irds.sparkler.SparklerConfig;
 import edu.usc.irds.sparkler.SparklerException;
 
 import javax.validation.ConstraintViolation;
