+++ sparkler-plugins/urlfilter-regex/src/test/java/edu/usc/irds/sparkler/config/UrlFilterPropsTest.java
@@ -1,7 +1,6 @@
 package edu.usc.irds.sparkler.config;
 
 import edu.usc.irds.sparkler.Constants;
-import edu.usc.irds.sparkler.SparklerConfig;
 import org.junit.Test;
 
 public class UrlFilterPropsTest {
